Conflicts:
	project4/project4/StatModule.cpp
	project4/project4/StatModule.h
	project4/project4/cpu.cpp
	project4/project4/cpu.h
	project4/project4/discrete_generator.cpp
	project4/project4/job.cpp
	project4/project4/job.h
	project4/project4/main.cpp

+++ project4/project4/job.cpp
@@ -57,12 +57,13 @@
 			else
 				return true;
 		}
+
 int Job::getProcTime()
 {
     return processTime;
 }
 
-int Job::getWait()
+double Job::getWait()
 {
     return waitTime;
 }
